Updated Common

+++ Data/Model/Ingredient.cs
@@ -11,10 +11,10 @@ namespace Godwit.Common.Data.Model {
         public string Name { get; set; }
         public MeasurementUnit MeasurementUnit { get; set; }
         public float ServingValue { get; set; }
-        public short CaloriesPerServing { get; set; }
-        public short CarbsPerServing { get; set; }
-        public short FatPerServing { get; set; }
-        public short ProteinPerServing { get; set; }
+        public float CaloriesPerServing { get; set; }
+        public float CarbsPerServing { get; set; }
+        public float FatPerServing { get; set; }
+        public float ProteinPerServing { get; set; }
         public Instant CreatedOn { get; set; }
         public Instant? UpdatedOn { get; set; }
         public virtual ICollection<Intake> Intakes { get; set; }
